api/requirements.txt
channels
daphne
channels-redis
django-channels-jwt-auth-middleware
django
pillow
djangorestframework
djangorestframework-simplejwt

